requirements.txt
pytest
beautifulsoup4
mysql-connector
Flask
httpie
flask-cors
textblob
nltk